+++ src/geometry/constrain.rs
@@ -7,6 +7,10 @@ pub enum Constrain {
     Percent(usize),
     /// Minimum size of the widget, widget expands to fit content
     Min(usize),
+    /// Maximum size of the widget, widget expands to fit content
+    Max(usize),
+    /// Minimum and maximum size of the widget, widget expands to fit content
+    MinMax(usize, usize),
     /// Fills rest of the space (space is divided by all widgets with fill)
     Fill,
 }