+++ src/main.rs
@@ -18,14 +18,15 @@ use termint::{
 
 fn main() {
     println!("\x1b[2J");
-    let mut grad = BgGrad::new((0, 150, 255), (150, 255, 0))
-        .center()
-        .bg_dir(Direction::Vertical);
-    let mut layout = Layout::horizontal().center();
-    layout.add_child(Block::new(), Constrain::Length(6));
-    grad.add_child(layout, Constrain::Length(3));
-    grad.render(&Coords::new(1, 1), &Coords::new(20, 9));
-    println!("\x1b[5B");
+
+    let mut layout = Block::new();
+    layout.add_child("Test", Constrain::Min(1));
+    layout.add_child("Testing of this really cool feature", Constrain::Max(3));
+    layout.add_child("Please just work", Constrain::MinMax(1, 3));
+    layout.render(&Coords::new(1, 1), &Coords::new(10, 8));
+
+    println!("\x1b[1B");
+
     // test_block();
     // test_layout();
     // test_grad();
@@ -224,4 +225,17 @@ fn test_layout_centering() {
     main.add_child(block, Constrain::Length(4));
     main.render(&Coords::new(1, 1), &Coords::new(20, 8));
     println!("\x1b[3B");
+}
+
+#[allow(unused)]
+fn test_bg_grad() {
+    println!("\x1b[2J");
+    let mut grad = BgGrad::new((0, 150, 255), (150, 255, 0))
+        .center()
+        .bg_dir(Direction::Vertical);
+    let mut layout = Layout::horizontal().center();
+    layout.add_child(Block::new(), Constrain::Length(6));
+    grad.add_child(layout, Constrain::Length(3));
+    grad.render(&Coords::new(1, 1), &Coords::new(20, 9));
+    println!("\x1b[5B");
 }